Improve backup fix script
The check-backup script now only checks if backup is valid
Add -f option to create a fixed backup
